corrections and new gaps merging plugin

+++ savu/plugins/segmentation/evolving_contours/mask_evolve3D.py
@@ -66,7 +66,7 @@ class MaskEvolve3d(Plugin, MultiThreadedPlugin):
         indices = np.where(np.isnan(input_temp))
         input_temp[indices] = 0.0
         if (np.sum(data[1]) > 0):
-            mask_evolve = MASK_ITERATE(input_temp, data[1], self.threshold, self.iterations, self.connectivity, self.method)
+            mask_evolve = MASK_ITERATE(input_temp, np.uint8(data[1]), self.threshold, self.iterations, self.connectivity, self.method)
         else:
             mask_evolve = np.uint8(np.zeros(np.shape(data[0])))
         return mask_evolve
